tests unitaires complétés

+++ database/migrations/2020_12_07_113947_create_avis_table.php
@@ -15,7 +15,7 @@ class CreateAvisTable extends Migration
     {
         Schema::create('avis', function (Blueprint $table) {
             $table->id();
-            $table->float('note');
+            $table->integer('note');
             $table->text('commentaire')->nullable();
             $table->timestamps();
